# apps/api-gateway/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: AP Calculus Bot API
  description: API Gateway for AP Calculus AB/BC tutoring bot with verified answers
  version: 0.1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /webhooks/stripe:
    post:
      tags:
        - webhooks
      description: Handle Stripe webhook events
      parameters:
        - schema:
            type: string
          in: header
          name: stripe-signature
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  message:
                    type: string
                required:
                  - success
                  - message
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
...
  /webhooks/stripe/config:
    get:
      tags:
        - webhooks
      description: Get Stripe webhook configuration
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    type: string
                  supportedEvents:
                    type: array
                    items:
                      type: string
                  signatureVerification:
                    type: boolean
                  idempotency:
                    type: boolean
                required:
                  - provider
                  - supportedEvents
                  - signatureVerification
                  - idempotency
                additionalProperties: false
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
servers:
  - url: http://localhost:3001
    description: Development server